// src/types/products.ts
export interface Product {
     id: number;
     slug: string;
     nameBn: string;
     category: string;
     categoryNameBn: string;
     categoryIcon: string;
     unit: string;
     image: string;
     today: number;
     yesterday: number;
     lastWeek: number;
     lastMonth: number;
     change: Change;
     markets: Market[];
}

export interface Change {
     dir: 'up' | 'down' | 'flat';
     pct: number;
}

export interface Market {
     market: string;
     division: string;
     min: number;
     max: number;
}